+++ src/components/2013/page-a-look-at-sparql/page-a-look-at-sparql.tsx
@@ -1,8 +1,8 @@
 import { Component, h } from '@stencil/core';
 
 import Prism from "prismjs"
-import '/assets/prismjs/prism-turtle.min.js';
-import '/assets/prismjs/prism-sparql.min.js';
+import 'prismjs/components/prism-turtle.min.js';
+import 'prismjs/components/prism-sparql.min.js';
 
 import { BlogData } from '../../../services/blog-data';