+++ src/components/2018/page-vscode-cheat-sheet/page-vscode-cheat-sheet.tsx
@@ -10,8 +10,8 @@ import { Component, h } from '@stencil/core';
 // import 'prismjs/components/prism-yaml.min';
 // import 'prismjs/components/prism-java.min';
 // These two both for SPARQL:
-// import '/assets/prismjs/prism-turtle.min.js';
-// import '/assets/prismjs/prism-sparql.min.js';
+// import 'prismjs/components/prism-turtle.min.js';
+// import 'prismjs/components/prism-sparql.min.js';
 
 import { BlogData } from '../../../services/blog-data';